feat(runtime): extract action-execution subsystem — ADR-0076 D11 step ③ PR-8 (#2462)
16 helpers (~560 lines) behind server-registered business actions move
to action-execution.ts on the narrow ActionExecutionDeps slice
(resolveService + getObjectQL — deliberately NO env-resolution state;
that stays with route handlers). The ADR-0104 warn-once statics
(actionParamsStrict/warnActionParamsOnce) ride along as module
functions — they were HttpDispatcher statics the extracted body
referenced, which would have made a value-cycle back into the
dispatcher.

Dispatcher keeps 4 thin delegates that have in-class callers
(callData/enforceActionParams/actionPermissionError/buildActionSession);
the 12 internal-only helpers are invoked directly on the module (5 call
sites in buildMcpBridge/handleActions rewired).

This is the pre-cut for PR-9: /actions and /mcp become mechanical
domain extractions now that their shared machinery is a module.

Verified: runtime 649, http-conformance 41, dogfood 351 (full-build
first — the suite needs every example/connector dist), dependent
closure DTS green.
